code style and refactoring
